sandbox/seccomp: move remaining seccomp bits from release
The patch moves seccomp related APIs from the release pacakge into the
sandbox/seccomp one.
