+++ sandbox/seccomp/seccomp_test.go
@@ -17,14 +17,14 @@
  *
  */
 
-package release_test
+package seccomp_test
 
 import (
 	"io"
 
 	. "gopkg.in/check.v1"
 
-	"github.com/snapcore/snapd/release"
+	"github.com/snapcore/snapd/sandbox/seccomp"
 )
 
 type seccompSuite struct{}
@@ -32,38 +32,38 @@ type seccompSuite struct{}
 var _ = Suite(&seccompSuite{})
 
 func (s *seccompSuite) TestInterfaceSystemKey(c *C) {
-	reset := release.MockSecCompActions([]string{})
+	reset := seccomp.MockActions([]string{})
 	defer reset()
-	c.Check(release.SecCompActions(), DeepEquals, []string{})
+	c.Check(seccomp.Actions(), DeepEquals, []string{})
 
-	reset = release.MockSecCompActions([]string{"allow", "errno", "kill", "log", "trace", "trap"})
+	reset = seccomp.MockActions([]string{"allow", "errno", "kill", "log", "trace", "trap"})
 	defer reset()
-	c.Check(release.SecCompActions(), DeepEquals, []string{"allow", "errno", "kill", "log", "trace", "trap"})
+	c.Check(seccomp.Actions(), DeepEquals, []string{"allow", "errno", "kill", "log", "trace", "trap"})
 }
 
 func (s *seccompSuite) TestSecCompSupportsAction(c *C) {
-	reset := release.MockSecCompActions([]string{})
+	reset := seccomp.MockActions([]string{})
 	defer reset()
-	c.Check(release.SecCompSupportsAction("log"), Equals, false)
+	c.Check(seccomp.SupportsAction("log"), Equals, false)
 
-	reset = release.MockSecCompActions([]string{"allow", "errno", "kill", "log", "trace", "trap"})
+	reset = seccomp.MockActions([]string{"allow", "errno", "kill", "log", "trace", "trap"})
 	defer reset()
-	c.Check(release.SecCompSupportsAction("log"), Equals, true)
+	c.Check(seccomp.SupportsAction("log"), Equals, true)
 }
 
 func (s *seccompSuite) TestProbe(c *C) {
-	release.FreshSecCompProbe()
-	r1 := release.MockIoutilReadfile(func(string) ([]byte, error) {
+	seccomp.FreshSecCompProbe()
+	r1 := seccomp.MockIoutilReadfile(func(string) ([]byte, error) {
 		return []byte("a b\n"), nil
 	})
 	defer r1()
 
-	c.Check(release.SecCompActions(), DeepEquals, []string{"a", "b"})
+	c.Check(seccomp.Actions(), DeepEquals, []string{"a", "b"})
 
-	r2 := release.MockIoutilReadfile(func(string) ([]byte, error) {
+	r2 := seccomp.MockIoutilReadfile(func(string) ([]byte, error) {
 		return nil, io.ErrUnexpectedEOF
 	})
 	defer r2()
 
-	c.Check(release.SecCompActions(), DeepEquals, []string{"a", "b"})
+	c.Check(seccomp.Actions(), DeepEquals, []string{"a", "b"})
 }
